Point, PointInt
* SDL.FPoint > SDL.Point
* SDL.Point > SDL.PointInt

+++ Engine/Framework/Internal/SDL3/SDL/Core/Point.cs
@@ -5,7 +5,7 @@ internal static unsafe partial class SDL
     [StructLayout(LayoutKind.Sequential)]
     public struct Point
     {
-        public int x;
-        public int y;
+        public float x;
+        public float y;
     }
 }
